Website: Remove unused React components
This is a cleanup of a few components we aren't using anymore (Select,
VersionSelect). Nothing fancy. Confirm the CI checks pass and merge.
