--- packages/playground/website/src/components/edit-site-settings-button/style.module.css
@@ -1,3 +1,0 @@
-.open-site-manager-button:global(> svg) {
-	margin-top: 4px;
-}
