--- packages/playground/website/src/components/select/style.module.css
@@ -1,13 +1,0 @@
-.select {
-	flex-grow: 1;
-	padding: 0px 10px;
-	font-family: 'San Francisco', Helvetica, Arial, sans-serif;
-	font-size: 16px;
-	font-weight: 50;
-	height: 26px;
-	border: 0;
-	background: #40464d;
-	border-radius: 8px;
-	color: #a5afbc;
-	transition: color 0.5s ease;
-}
